+++ src/.internal/core/Registry.ts
@@ -6,7 +6,7 @@ export class Registry {
 	/**
 	 * Currently running version of amCharts.
 	 */
-	readonly version: string = "5.4.11";
+	readonly version: string = "5.5.0";
 
 	/**
 	 * List of applied licenses.